requirements.txt
attrs==22.2.0
certifi==2022.12.7
charset-normalizer==3.0.1
decorator==5.1.1
deepdiff==6.2.3
idna==3.4
jsonpath-ng==1.5.3
jsonschema==4.17.3
ordered-set==4.1.0
orjson==3.8.6
ply==3.11
pyrsistent==0.19.3
python-dotenv==0.21.1
requests==2.28.2
robotframework==6.0.2
robotframework-jsonlibrary==0.5
six==1.16.0
urllib3==1.26.14
